Refactored dialog for creating events ready for editing. Relates to lewisoaten/caLANdar#48-edit-events
